% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDocByCnes}
\alias{getDocByCnes}
\title{Obtem todo o documento mongodb relativo a um dado numero CNES}
\usage{
getDocByCnes(cnesCod, ns = "opensus.cnesminas")
}
\arguments{
\item{cnesCod}{Codigo CNES a ser pesquisado}
}
\description{
Obtem todo o documento mongodb relativo a um dado numero CNES
}
\author{
Jose Moraes
}

