% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDocByFieldsList}
\alias{getDocByFieldsList}
\title{Get Fields values based a R named list in te form fieldsList = ("Key" = "value" )}
\usage{
getDocByFieldsList(fieldsList, ns = "opensus.cnesminas", nLimit = 10L)
}
\arguments{
\item{fieldsList}{a R named list containing ("key" = "VALUE") pairs}

\item{ns}{}

\item{nLimit}{}
}
\description{
Get Fields values based a R named list in te form fieldsList = ("Key" = "value" )
}
\author{
Jose Moraes
}

