% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDocByMunIBGECod}
\alias{getDocByMunIBGECod}
\title{Get a \code{R} \code{list} containing all documents in the \code{mongodb}
for a municipio given its IBGE number code ID}
\usage{
getDocByMunIBGECod(munIBGECod)
}
\arguments{
\item{munIBGECod}{}
}
\description{
Get a \code{R} \code{list} containing all documents in the \code{mongodb}
for a municipio given its IBGE number code ID
}

