% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDocByMunicipio}
\alias{getDocByMunicipio}
\title{Obtem um objeto da classe \code{R List} com todos os
campos existentes no documentos \code{mongodb}}
\usage{
getDocByMunicipio(municipio, ns = "opensus.cnesminas", nLimite = 10L)
}
\arguments{
\item{ns}{\code{namespace} na instancia \code{mongodb} a ser usado na pesquisa}

\item{nLimite}{numero maximo de documentos a ser retornado, padrao 10.}

\item{cnesCod}{Codigo CNES a ser pesquisado}
}
\description{
Obtem um objeto da classe \code{R List} com todos os
campos existentes no documentos \code{mongodb}
}
\author{
Jose Moraes
}

