% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDocStructure}
\alias{getDocStructure}
\title{Convenience method to print the fields contained in \code{mongodb document}
as a dataframe. Namespace is hard coded as \code{ns = "opensus.cnesminas"}}
\usage{
getDocStructure()
}
\value{
\code{mongodb document}
}
\description{
Convenience method to print the fields contained in \code{mongodb document}
as a dataframe. Namespace is hard coded as \code{ns = "opensus.cnesminas"}
}
\author{
Jose Moraes
}

