% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getGeoJSONByMunicipio}
\alias{getGeoJSONByMunicipio}
\title{geoJSON representation of a given municipio \code{string} name and a
\code{integer} to limit number os records returned}
\usage{
getGeoJSONByMunicipio(municipio = "Vargem Alegre", ns = "opensus.cnesminas",
  nLimit = 10L)
}
\arguments{
\item{municipio}{municipio name in \code{string} format}

\item{ns}{namespace of \code{mongodb} instance default: "opensus.cnesminas"}

\item{nLimit}{maximum number of records to br returned}
}
\description{
geoJSON representation of a given municipio \code{string} name and a
\code{integer} to limit number os records returned
}
\author{
Jose Moraes
}

