% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLatLonByCnes}
\alias{getLatLonByCnes}
\title{Retrive a dataframe containing Latitud and Longitude for a given CNES cod}
\usage{
getLatLonByCnes(cnesCod, ns = "opensus.cnesminas")
}
\arguments{
\item{cnesCod}{CNES number of a institution}
}
\value{
res a data frame with values for Lat and Lon
}
\description{
Retrive a dataframe containing Latitud and Longitude for a given CNES cod
}
\author{
Jose Moraes
}

