% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getLatLonByMunicipio}
\alias{getLatLonByMunicipio}
\title{Retrieve a data.frame containing all Latitude and Longitude
values for a given municipio name}
\usage{
getLatLonByMunicipio(municipio, ns = "opensus.cnesminas", nLimit = 30000L)
}
\arguments{
\item{municipio}{name of the municipio}
}
\value{
res a data.frame with values for Lat and Lon
}
\description{
Retrieve a data.frame containing all Latitude and Longitude
values for a given municipio name
}
\author{
Jose Moraes
}

