% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMapByLatLon}
\alias{getMapByLatLon}
\title{Get a map based on a pair of \code{\[lat, lon\]} boundingbox coord passed}
\usage{
getMapByLatLon(ULClat, ULClon, BRClat, BRClon)
}
\arguments{
\item{ULClat}{}

\item{ULClon}{}

\item{BRClat}{}

\item{BRClon}{}
}
\value{
a map
}
\description{
Get a map based on a pair of \code{\[lat, lon\]} boundingbox coord passed
}
\author{
Jose Moraes
}

