% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMapByMunIBGECod}
\alias{getMapByMunIBGECod}
\title{Get a \code{ggmap} \code{map} for a municipio given its IBGE number code ID}
\usage{
getMapByMunIBGECod(munIBGECod)
}
\arguments{
\item{munIBGECod}{}
}
\description{
Get a \code{ggmap} \code{map} for a municipio given its IBGE number code ID
}

