% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getMapByMunicipioList}
\alias{getMapByMunicipioList}
\title{Get Fields values based a R named list in te form fieldsList = ("Key" = "value" )}
\usage{
getMapByMunicipioList(municipio, mapsource = "osm",
  ns = "opensus.cnesminas", nLimit = 10L)
}
\arguments{
\item{ns}{}

\item{nLimit}{}

\item{fieldsList}{a R named list containing ("key" = "VALUE") pairs}
}
\description{
Get Fields values based a R named list in te form fieldsList = ("Key" = "value" )
}
\author{
Jose Moraes
}

