% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getNumDocByMunicipio}
\alias{getNumDocByMunicipio}
\title{Obtem o numero de resistros (documentos na instancia do mongodb)}
\usage{
getNumDocByMunicipio(municipio, ns = "opensus.cnesminas")
}
\arguments{
\item{municipio}{nome do municipio}

\item{ns}{namespace a ser usado, na forma 'database.collection'}
}
\description{
Obtem o numero de resistros (documentos na instancia do mongodb)
}
\author{
Jose Moraes
}

