% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getDocByCnes}
\alias{getDocByCnes}
\title{Retrieve the \code{mongodb document} given a CNES code number}
\usage{
getDocByCnes(cnesCod, ns = "opensus.cnesminas")
}
\arguments{
\item{cnesCod}{Codigo CNES a ser pesquisado}

\item{ns}{\code{mongodb} namespace}
}
\description{
Retrieve the \code{mongodb document} given a CNES code number
}
\author{
Jose Moraes
}

